Collapsed document-tab-content and document-tab-panel into a single component called section-document-or-browser
