+++ src/components/thumbnail/tab-panel-documents-subsection-panel.tsx
@@ -5,7 +5,7 @@ import { useFirestoreTeacher } from "../../hooks/firestore-hooks";
 import { DocumentModelType } from "../../models/document/document";
 import { isPublishedType, SupportPublication } from "../../models/document/document-types";
 import { getDocumentDisplayTitle } from "../../models/document/document-utils";
-import { IStores } from "../../models/stores/stores";
+import { useAppConfig, useClassStore, useProblemStore, useUserStore } from "../../hooks/use-stores";
 import { NavTabSectionModelType } from "../../models/view/nav-tabs";
 
 import "./tab-panel-documents-section.sass";
@@ -14,7 +14,6 @@ interface IProps {
   sectionDocument: DocumentModelType;
   tab: string;
   section: NavTabSectionModelType;
-  stores: IStores;
   scale: number;
   selectedDocument?: string;
   onSelectDocument?: (document: DocumentModelType) => void;
@@ -23,12 +22,15 @@ interface IProps {
   onDocumentDeleteClick?: (document: DocumentModelType) => void;
 }
 
-function useDocumentCaption(stores: IStores, document: DocumentModelType) {
-  const { appConfig, problem, class: _class, user: { network } } = stores;
+function useDocumentCaption(document: DocumentModelType) {
+  const appConfig = useAppConfig();
+  const problem = useProblemStore();
+  const classStore = useClassStore();
+  const user = useUserStore();
   const { type, uid } = document;
-  const teacher = useFirestoreTeacher(uid, network || "");
+  const teacher = useFirestoreTeacher(uid, user.network || "");
   if (type === SupportPublication) return document.getProperty("caption") || "Support";
-  const userName = _class?.getUserById(uid)?.displayName ||
+  const userName = classStore.getUserById(uid)?.displayName ||
                     (document.isRemote ? teacher?.name : "") || "Unknown User";
   const namePrefix = document.isRemote || isPublishedType(type) ? `${userName}: ` : "";
   const dateSuffix = document.isRemote && document.createdAt
@@ -39,12 +41,12 @@ function useDocumentCaption(stores: IStores, document: DocumentModelType) {
 
 // observes teacher names via useDocumentCaption()
 export const TabPanelDocumentsSubSectionPanel = observer(({
-  section, sectionDocument, tab, stores, scale, selectedDocument,
+  section, sectionDocument, tab, scale, selectedDocument,
   onSelectDocument, onDocumentDragStart, onDocumentStarClick, onDocumentDeleteClick
 }: IProps) => {
-    const { user } = stores;
+    const user = useUserStore();
     const tabName = tab.toLowerCase().replace(' ', '-');
-    const caption = useDocumentCaption(stores, sectionDocument);
+    const caption = useDocumentCaption(sectionDocument);
 
     function handleDocumentClick() {
       onSelectDocument?.(sectionDocument);
